requirements.txt
pandas
seaborn
matplotlib
numpy
scikit-learn
tensorflow
azureml-core
azureml-sdk
joblib